Augment an mvgam object's data

Add fits and residuals to the data, implementing the generic augment from the package broom.

Usage

# S3 method for class 'mvgam'
augment(x, robust = FALSE, probs = c(0.025, 0.975), ...)

Arguments

x

An object of class mvgam.

robust

If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead.

probs

The percentiles to be computed by the quantile function.

...

Unused, included for generic consistency only.

Value

A list or tibble (see details) combining:

The data supplied to mvgam().

The outcome variable, named as .observed.

The fitted backcasts, along with their variability and credible bounds.

The residuals, along with their variability and credible bounds.

Details

A list is returned if class(x$obs_data) == 'list', otherwise a tibble is returned, but the contents of either object is the same.

The arguments robust and probs are applied to both the fit and residuals calls (see fitted.mvgam() and residuals.mvgam() for details).

See also

Examples

# \donttest{
set.seed(0)
dat <- sim_mvgam(T = 80,
                 n_series = 3,
                 mu = 2,
                 trend_model = AR(p = 1),
                 prop_missing = 0.1,
                 prop_trend = 0.6)

mod1 <- mvgam(formula = y ~ s(season, bs = 'cc', k = 6),
              data = dat$data_train,
              trend_model = AR(),
              family = poisson(),
              noncentred = TRUE,
              chains = 2,
              silent = 2)

augment(mod1, robust = TRUE, probs = c(0.25, 0.75))
#> # A tibble: 180 × 14
#>        y season  year series    time .observed .fitted .fit.variability
#>    <int>  <int> <int> <fct>    <int>     <int>   <dbl>            <dbl>
#>  1     4      1     1 series_1     1         4    4.49             2.95
#>  2    NA      1     1 series_2     1        NA    6.13             3.68
#>  3     4      1     1 series_3     1         4    4.35             3.02
#>  4     5      2     1 series_1     2         5    4.51             2.78
#>  5     2      2     1 series_2     2         2    3.61             3.45
#>  6    NA      2     1 series_3     2        NA    4.52             3.68
#>  7     7      3     1 series_1     3         7    8.38             4.05
#>  8    12      3     1 series_2     3        12   11.3              5.61
#>  9     4      3     1 series_3     3         4    5.13             3.05
#> 10    39      4     1 series_1     4        39   35.8             27.9 
#> # ℹ 170 more rows
#> # ℹ 6 more variables: .fit.cred.low <dbl>, .fit.cred.high <dbl>, .resid <dbl>,
#> #   .resid.variability <dbl>, .resid.cred.low <dbl>, .resid.cred.high <dbl>
# }