Extract or compute hindcasts and forecasts for a fitted mvgam object

Extract or compute hindcasts and forecasts for a fitted mvgam object

Usage

forecast(object, ...)

# S3 method for mvgam
forecast(
  object,
  newdata,
  data_test,
  series = "all",
  n_cores = 1,
  type = "response",
  ...
)

Arguments

object

list object returned from mvgam. See mvgam()

...

Ignored

newdata

Optional dataframe or list of test data containing at least 'series' and 'time' in addition to any other variables included in the linear predictor of the original formula. If included, the covariate information in newdata will be used to generate forecasts from the fitted model equations. If this same newdata was originally included in the call to mvgam, then forecasts have already been produced by the generative model and these will simply be extracted and plotted. However if no newdata was supplied to the original model call, an assumption is made that the newdata supplied here comes sequentially after the data supplied as data in the original model (i.e. we assume there is no time gap between the last observation of series 1 in data and the first observation for series 1 in newdata)

data_test

Deprecated. Still works in place of newdata but users are recommended to use newdata instead for more seamless integration into R workflows

series

Either a integer specifying which series in the set is to be forecast, or the character string 'all', specifying that all series should be forecast. This is preferable if the fitted model contained multivariate trends (either as a dynamic factor or VAR process), as it saves recomputing the full set of trends for each series individually

n_cores

integer specifying number of cores for generating forecasts in parallel

type

When this has the value link (default) the linear predictor is calculated on the link scale. If expected is used, predictions reflect the expectation of the response (the mean) but ignore uncertainty in the observation process. When response is used, the predictions take uncertainty in the observation process into account to return predictions on the outcome scale. When trend is used, only the forecast distribution for the latent trend is returned

Value

An object of class mvgam_forecast containing hindcast and forecast distributions. See mvgam_forecast-class for details.

Details

Posterior predictions are drawn from the fitted mvgam and used to simulate a forecast distribution