Extract information on default prior distributions for an mvgam model

This function lists the parameters that can have their prior distributions changed for a given mvgam model, as well listing their default distributions

Usage

get_mvgam_priors(
  formula,
  data,
  data_train,
  family = "poisson",
  use_lv = FALSE,
  n_lv,
  use_stan = TRUE,
  trend_model = "None",
  drift = FALSE
)

Arguments

formula

A character string specifying the GAM formula. These are exactly like the formula for a GLM except that smooth terms, s, te, ti and t2, can be added to the right hand side to specify that the linear predictor depends on smooth functions of predictors (or linear functionals of these)

data

A dataframe or list containing the model response variable and covariates required by the GAM formula. Should include columns: 'y' (the discrete outcomes; NAs allowed) 'series' (character or factor index of the series IDs) 'time' (numeric index of the time point for each observation). Any other variables to be included in the linear predictor of formula must also be present

data_train

Deprecated. Still works in place of data but users are recommended to use data instead for more seamless integration into R workflows

family

character. Must be either 'nb' (for Negative Binomial), 'tw' (for Tweedie) or 'poisson'

use_lv

logical. If TRUE, use dynamic factors to estimate series' latent trends in a reduced dimension format. If FALSE, estimate independent latent trends for each series

n_lv

integer the number of latent dynamic factors to use if use_lv == TRUE. Cannot be >n_series. Defaults arbitrarily to min(2, floor(n_series / 2))

use_stan

Logical. If TRUE and if rstan is installed, the model will be compiled and sampled using the Hamiltonian Monte Carlo with a call to cmdstan_model or, if cmdstanr is not available, a call to stan. Note that this functionality is still in development and not all options that are available in JAGS can be used, including: no option for a Tweedie family and no option for dynamic factor trends. However, as Stan can estimate Hilbert base approximate Gaussian Processes, which are much more computationally tractable than full GPs for time series with >100 observations, estimation in Stan can support latent GP trends while estimation in JAGS cannot

trend_model

character specifying the time series dynamics for the latent trend. Options are: 'None' (no latent trend component; i.e. the GAM component is all that contributes to the linear predictor, and the observation process is the only source of error; similarly to what is estimated by gam), 'RW' (random walk with possible drift), 'AR1' (AR1 model with intercept), 'AR2' (AR2 model with intercept) or 'AR3' (AR3 model with intercept) or 'VAR1' (with possible drift; only available in Stan) or 'GP' (Gaussian Process with squared exponential kernel; only available for estimation in Stan)

drift

logical estimate a drift parameter in the latent trend components. Useful if the latent trend is expected to broadly follow a non-zero slope. Note that if the latent trend is more or less stationary, the drift parameter can become unidentifiable, especially if an intercept term is included in the GAM linear predictor (which it is by default when calling jagam). Therefore this defaults to FALSE

Value

either a data.frame containing the prior definitions (if any suitable priors can be altered by the user) or NULL, indicating that no priors in the model can be modified through the mvgam interface

Details

Users can supply a model formula, prior to fitting the model, so that default priors can be inspected and altered. To make alterations, change the contents of the prior column and supplying this data.frame to the mvgam function using the argument priors. If using Stan as the backend, users can also modify the parameter bounds by modifying the new_lowerbound and/or new_upperbound columns. This will be necessary if using restrictive distributions on some parameters, such as a Beta distribution for the trend sd parameters for example (Beta only has support on (0,1)), so the upperbound cannot be above 1

Note

Only the prior, new_lowerbound and/or new_upperbound columns of the output should be altered when defining the user-defined priors for the mvgam model. Use only if you are familiar with the underlying probabilistic programming language. There are no sanity checks done to ensure that the code is legal (i.e. to check that lower bounds are smaller than upper bounds, for example)

See also

Author

Nicholas J Clark

Examples

# Simulate three integer-valued time series
library(mvgam)
dat <- sim_mvgam(trend_rel = 0.5)

# Get a model file that uses default mvgam priors for inspection (not always necessary,
# but this can be useful for testing whether your updated priors are written correctly)
mod_default <- mvgam(y ~ s(series, bs = 're') +
              s(season, bs = 'cc') - 1,
              family = 'nb',
              data = dat$data_train,
              trend_model = 'AR2',
              priors = test_priors,
              run_model = FALSE)
#> Error in update_priors(vectorised$model_file, priors, use_stan = TRUE): object 'test_priors' not found

# Inspect the model file with default mvgam priors
mod_default$model_file
#> Error in eval(expr, envir, enclos): object 'mod_default' not found

# Look at which priors can be updated in mvgam
test_priors <- get_mvgam_priors(y ~ s(series, bs = 're') +
                                s(season, bs = 'cc') - 1,
                                family = 'nb',
                                data = dat$data_train,
                                trend_model = 'AR2')
test_priors
#>                                    param_name param_length
#> 1               vector<lower=0>[n_sp] lambda;            2
#> 2                           vector[1] mu_raw;            1
#> 3               vector<lower=0>[1] sigma_raw;            1
#> 4 vector<lower=-1.5,upper=1.5>[n_series] ar1;            3
#> 5 vector<lower=-1.5,upper=1.5>[n_series] ar2;            3
#> 6            vector<lower=0>[n_series] sigma;            3
#> 7          vector<lower=0>[n_series] phi_inv;            3
#>                    param_info                           prior
#> 1 s(season) smooth parameters        lambda ~ normal(10, 25);
#> 2          s(series) pop mean          mu_raw ~ std_normal();
#> 3            s(series) pop sd   sigma_raw ~ exponential(0.5);
#> 4       trend AR1 coefficient             ar1 ~ std_normal();
#> 5       trend AR2 coefficient             ar2 ~ std_normal();
#> 6                    trend sd         sigma ~ exponential(2);
#> 7   inverse of NB dispsersion phi_inv ~ student_t(3, 0, 0.1);
#>                   example_change new_lowerbound new_upperbound
#> 1    lambda ~ exponential(0.05);             NA             NA
#> 2  mu_raw ~ normal(-0.44, 0.11);             NA             NA
#> 3 sigma_raw ~ exponential(0.93);             NA             NA
#> 4      ar1 ~ normal(0.27, 0.26);             NA             NA
#> 5      ar2 ~ normal(-0.1, 0.63);             NA             NA
#> 6     sigma ~ exponential(0.03);             NA             NA
#> 7  phi_inv ~ normal(0.08, 0.48);             NA             NA

# Make a few changes; first, change the population mean for the series-level
# random intercepts
test_priors$prior[1] <- 'mu_raw1 ~ dnorm(-1, 2)'

# Now use stronger regularisation for the series-level AR2 coefficients
test_priors$prior[4] <- 'ar2[s] ~ dnorm(0, 20)'

# Check whether the notation in the 'prior' column is correct (note, this function
# will not check that densities are correctly spelled and parameterised, that is up
# to the user!)
update_priors(model_file = mod_default$model_file,
              priors = test_priors)
#> Error in update_priors(model_file = mod_default$model_file, priors = test_priors): could not find function "update_priors"

# No warnings, the model is ready for fitting now in the usual way with the addition
# of the 'priors' argument:
# mod <- mvgam(y ~ s(series, bs = 're') +
#                s(season, bs = 'cc') - 1,
#              family = 'nb',
#              data = dat$data_train,
#              trend_model = 'AR2',
#              priors = test_priors)

# Look at what is returned when an incorrect spelling is used
test_priors$prior[4] <- 'ar2_bananas ~ dnorm(0, 20)'
update_priors(model_file = mod_default$model_file,
              priors = test_priors)
#> Error in update_priors(model_file = mod_default$model_file, priors = test_priors): could not find function "update_priors"