mvgam: Multivariate (Dynamic) Generalized Additive Models

Fit Bayesian Dynamic Generalized Additive Models to multivariate observations. Users can build nonlinear State-Space models that can incorporate semiparametric effects in observation and process components, using a wide range of observation families. Estimation is performed using Markov Chain Monte Carlo with Hamiltonian Monte Carlo in the software 'Stan'. References: Clark & Wells (2022) doi:10.1111/2041-210X.13974 .

See also

Author

Maintainer: Nicholas J Clark nicholas.j.clark1214@gmail.com (ORCID)

Other contributors:

Sarah Heaps (ORCID) (VARMA parameterisations) [contributor]

Scott Pease (ORCID) (broom enhancements) [contributor]

Matthijs Hollanders (ORCID) (ggplot visualizations) [contributor]