mvgam_fevd object description

A mvgam_fevd object returned by function fevd. Run methods(class = "mvgam_fevd") to see an overview of available methods.

Details

A mvgam_fevd object contains a list of posterior forecast error variance decompositions, each stored as its own list

See also

Author

Nicholas J Clark