Forecast from a particle filtered mvgam object

This function generates a forecast from a set of particles that each capture a unique proposal about the current state of the system that was modelled in the mvgam object. The covariate and timepoint information from data_test is used to generate the GAM component forecast, while the trends are run forward in time according to their state space dynamics. The forecast is a weighted ensemble, with weights determined by each particle's proposal likelihood prior to the most recent assimilation step

Usage

pfilter_mvgam_fc(
  file_path = "pfilter",
  n_cores = 2,
  newdata,
  data_test,
  plot_legend = TRUE,
  legend_position = "topleft",
  ylim,
  return_forecasts = FALSE
)

Arguments

file_path

character string specifying the file path where the particles have been saved

n_cores

integer specifying number of cores for generating particle forecasts in parallel

newdata

A dataframe or list of test data containing at least 'series' and time', in addition to any other variables included in the linear predictor of formula

data_test

Deprecated. Still works in place of newdata but users are recommended to use newdata instead for more seamless integration into R workflows

plot_legend

logical stating whether to include a legend to highlight which observations were used for calibration and which were assimilated by the particle filter

legend_position

The legend location may be specified by setting x to a single keyword from the list "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". This places the legend on the inside of the plot frame at the given location.

ylim

Optional vector of y-axis limits (min, max). The same limits will be used for all plots

return_forecasts

logical. If TRUE, the returned list object will contain plots of forecasts as well as the forecast objects (each as a matrix of dimension n_particles x horizon)

Value

A named list containing functions that call base R plots of each series' forecast. Optionally the actual forecasts are returned within the list as a separate list of matrices