Automatic online particle filtering for assimilating new observations into a fitted mvgam model

This function operates sequentially on new observations in data_assim to update the posterior forecast distribution. It is a wrapper that calls pfilter_mvgam_smooth. In each iteration, the next observation is assimilated and particles are weighted by their proposal's multivariate composite likelihood

Usage

pfilter_mvgam_online(
  newdata,
  data_assim,
  file_path = "pfilter",
  threshold = 0.5,
  use_resampling = FALSE,
  kernel_lambda = 0.25,
  n_cores = parallel::detectCores() - 1
)

Arguments

newdata

A dataframe or list of test data containing at least one more observation per series (beyond the last observation seen by the model when initialising particles with pfilter_mvgam_init or in previous calls to pfilter_mvgam_online. Should at least contain 'series' and 'time' for the one-step ahead horizon, in addition to any other variables included in the linear predictor of object

data_assim

Deprecated. Still works in place of newdata but users are recommended to use newdata instead for more seamless integration into R workflows

file_path

character string specifying the file path for locating the particles

threshold

proportional numeric specifying the Effective Sample Size limit under which resampling of particles will be triggered (calculated as ESS / n_particles) if use_resampling == TRUE. Should be between 0 and 1

use_resampling

logical specifying whether resampling should be used when ESS falls below the specified threshold. Default for this option is FALSE, relying instead on kernel smoothing only to maintain particle diversity

kernel_lambda

proportional numeric specifying the strength of kernel smoothing to use when pulling low weight particles toward the high likelihood state space. Should be between 0 and 1

n_cores

integer specifying number of cores for generating particle forecasts in parallel

Value

A list object of length = n_particles containing information on parameters and current state estimates for each particle is generated and saved, along with other important information from the original model, to an .rda object in file_path