Assimilate new observations into a fitted mvgam model using resampling and kernel smoothing

This function operates on a new observation in next_assim to update the posterior forecast distribution. The next observation is assimilated and particle weights are updated in light of their most recent their multivariate composite likelihood. Low weight particles are smoothed towards the high weight state space using importance sampling, and options are given for using resampling of high weight particles when Effective Sample Size falls below a user-specified threshold

Usage

pfilter_mvgam_smooth(
  particles,
  mgcv_model,
  next_assim,
  threshold = 0.25,
  n_cores = 1,
  use_resampling = FALSE,
  kernel_lambda = 0.5
)

Arguments

particles

A list of particles that have been run up to one observation prior to the observation in next_assim

mgcv_model

A gam model returned through a call to link{mvgam}

next_assim

A dataframe of test data containing at one more observation per series (beyond the last observation seen by the model when initialising particles with pfilter_mvgam_init or in previous calls to pfilter_mvgam_online. Should at least contain 'series' and 'time' for the one-step ahead horizon, in addition to any other variables included in the linear predictor of object

threshold

proportional numeric specifying the Effective Sample Size limit under which resampling of particles will be triggered (calculated as ESS / n_particles) if use_resampling == TRUE. Should be between 0 and 1

n_cores

integer specifying number of cores for generating particle forecasts in parallel

use_resampling

logical specifying whether resampling should be used when ESS falls below the specified threshold. Note that resampling can result in loss of the original model's diversity of GAM beta coefficients, which may have undesirable consequences for the forecast distribution. If use_resampling is TRUE, some effort is made to remedy this by assigning randomly sampled draws of GAM beta coefficients from the original model's distribution to each particle. This does not however guarantee there will be no loss of diversity, especially if successive resampling take place. Default for this option is therefore FALSE

kernel_lambda

proportional numeric specifying the strength of smoothing to use when pulling low weight particles toward the high likelihood state space. Should be between 0 and 1

Value

A list object of length = n_particles containing information on parameters and current state estimates for each particle