Plot Pareto-k and ELPD values from a mvgam_lfo object

This function takes an object of class mvgam_lfo and creates several informative diagnostic plots

Usage

# S3 method for class 'mvgam_lfo'
plot(x, ...)

Arguments

x

An object of class mvgam_lfo

...

Ignored

Value

A ggplot object of Pareto-k and ELPD values over the evaluation timepoints. For the Pareto-k plot, a dashed red line indicates the specified threshold chosen for triggering model refits. For the ELPD plot, a dashed red line indicates the bottom 10% quantile of ELPD values. Points below this threshold may represent outliers that were more difficult to forecast