This function converts univariate or multivariate time series (xts or ts objects) to the format necessary for mvgam

This function converts univariate or multivariate time series (xts or ts objects) to the format necessary for mvgam

Usage

series_to_mvgam(series, freq, train_prop = 0.85)

Arguments

series

xts or ts object to be converted to mvgam format

freq

integer. The seasonal frequency of the series

train_prop

numeric stating the proportion of data to use for training. Should be between 0.25 and 0.95

Value

A list object containing outputs needed for mvgam, including 'data_train' and 'data_test'

Examples

# A ts object example
library(forecast)
#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 
#> 
#> Attaching package: 'forecast'
#> The following object is masked from 'package:mvgam':
#> 
#>     forecast
#> The following object is masked from 'package:nlme':
#> 
#>     getResponse
data("AirPassengers")
series <- cbind(AirPassengers, AirPassengers)
colnames(series) <- c('blood', 'bone')
series_to_mvgam(series, frequency(series), 0.85)
#> $data_train
#>       y season year                date series time
#> 1   112      1 1949 1949-01-01 00:00:00  blood    1
#> 2   112      1 1949 1949-01-01 00:00:00   bone    1
#> 3   118      2 1949 1949-01-31 10:00:00  blood    2
#> 4   118      2 1949 1949-01-31 10:00:00   bone    2
#> 5   132      3 1949 1949-03-02 20:00:01  blood    3
#> 6   132      3 1949 1949-03-02 20:00:01   bone    3
#> 7   129      4 1949 1949-04-02 06:00:00  blood    4
#> 8   129      4 1949 1949-04-02 06:00:00   bone    4
#> 9   121      5 1949 1949-05-02 16:00:00  blood    5
#> 10  121      5 1949 1949-05-02 16:00:00   bone    5
#> 11  135      6 1949 1949-06-02 02:00:01  blood    6
#> 12  135      6 1949 1949-06-02 02:00:01   bone    6
#> 13  148      7 1949 1949-07-02 12:00:00  blood    7
#> 14  148      7 1949 1949-07-02 12:00:00   bone    7
#> 15  148      8 1949 1949-08-01 22:00:00  blood    8
#> 16  148      8 1949 1949-08-01 22:00:00   bone    8
#> 17  136      9 1949 1949-09-01 08:00:01  blood    9
#> 18  136      9 1949 1949-09-01 08:00:01   bone    9
#> 19  119     10 1949 1949-10-01 18:00:00  blood   10
#> 20  119     10 1949 1949-10-01 18:00:00   bone   10
#> 21  104     11 1949 1949-11-01 04:00:00  blood   11
#> 22  104     11 1949 1949-11-01 04:00:00   bone   11
#> 23  118     12 1949 1949-12-01 14:00:01  blood   12
#> 24  118     12 1949 1949-12-01 14:00:01   bone   12
#> 25  115      1 1950 1950-01-01 00:00:00  blood   13
#> 26  115      1 1950 1950-01-01 00:00:00   bone   13
#> 27  126      2 1950 1950-01-31 10:00:00  blood   14
#> 28  126      2 1950 1950-01-31 10:00:00   bone   14
#> 29  141      3 1950 1950-03-02 20:00:01  blood   15
#> 30  141      3 1950 1950-03-02 20:00:01   bone   15
#> 31  135      4 1950 1950-04-02 06:00:00  blood   16
#> 32  135      4 1950 1950-04-02 06:00:00   bone   16
#> 33  125      5 1950 1950-05-02 16:00:00  blood   17
#> 34  125      5 1950 1950-05-02 16:00:00   bone   17
#> 35  149      6 1950 1950-06-02 02:00:01  blood   18
#> 36  149      6 1950 1950-06-02 02:00:01   bone   18
#> 37  170      7 1950 1950-07-02 12:00:00  blood   19
#> 38  170      7 1950 1950-07-02 12:00:00   bone   19
#> 39  170      8 1950 1950-08-01 22:00:00  blood   20
#> 40  170      8 1950 1950-08-01 22:00:00   bone   20
#> 41  158      9 1950 1950-09-01 08:00:01  blood   21
#> 42  158      9 1950 1950-09-01 08:00:01   bone   21
#> 43  133     10 1950 1950-10-01 18:00:00  blood   22
#> 44  133     10 1950 1950-10-01 18:00:00   bone   22
#> 45  114     11 1950 1950-11-01 04:00:00  blood   23
#> 46  114     11 1950 1950-11-01 04:00:00   bone   23
#> 47  140     12 1950 1950-12-01 14:00:01  blood   24
#> 48  140     12 1950 1950-12-01 14:00:01   bone   24
#> 49  145      1 1951 1951-01-01 00:00:00  blood   25
#> 50  145      1 1951 1951-01-01 00:00:00   bone   25
#> 51  150      2 1951 1951-01-31 10:00:00  blood   26
#> 52  150      2 1951 1951-01-31 10:00:00   bone   26
#> 53  178      3 1951 1951-03-02 20:00:01  blood   27
#> 54  178      3 1951 1951-03-02 20:00:01   bone   27
#> 55  163      4 1951 1951-04-02 06:00:00  blood   28
#> 56  163      4 1951 1951-04-02 06:00:00   bone   28
#> 57  172      5 1951 1951-05-02 16:00:00  blood   29
#> 58  172      5 1951 1951-05-02 16:00:00   bone   29
#> 59  178      6 1951 1951-06-02 02:00:01  blood   30
#> 60  178      6 1951 1951-06-02 02:00:01   bone   30
#> 61  199      7 1951 1951-07-02 12:00:00  blood   31
#> 62  199      7 1951 1951-07-02 12:00:00   bone   31
#> 63  199      8 1951 1951-08-01 22:00:00  blood   32
#> 64  199      8 1951 1951-08-01 22:00:00   bone   32
#> 65  184      9 1951 1951-09-01 08:00:01  blood   33
#> 66  184      9 1951 1951-09-01 08:00:01   bone   33
#> 67  162     10 1951 1951-10-01 18:00:00  blood   34
#> 68  162     10 1951 1951-10-01 18:00:00   bone   34
#> 69  146     11 1951 1951-11-01 04:00:00  blood   35
#> 70  146     11 1951 1951-11-01 04:00:00   bone   35
#> 71  166     12 1951 1951-12-01 14:00:01  blood   36
#> 72  166     12 1951 1951-12-01 14:00:01   bone   36
#> 73  171      1 1952 1952-01-01 00:00:00  blood   37
#> 74  171      1 1952 1952-01-01 00:00:00   bone   37
#> 75  180      2 1952 1952-01-31 12:00:00  blood   38
#> 76  180      2 1952 1952-01-31 12:00:00   bone   38
#> 77  193      3 1952 1952-03-02 00:00:01  blood   39
#> 78  193      3 1952 1952-03-02 00:00:01   bone   39
#> 79  181      4 1952 1952-04-01 12:00:00  blood   40
#> 80  181      4 1952 1952-04-01 12:00:00   bone   40
#> 81  183      5 1952 1952-05-02 00:00:00  blood   41
#> 82  183      5 1952 1952-05-02 00:00:00   bone   41
#> 83  218      6 1952 1952-06-01 12:00:01  blood   42
#> 84  218      6 1952 1952-06-01 12:00:01   bone   42
#> 85  230      7 1952 1952-07-02 00:00:00  blood   43
#> 86  230      7 1952 1952-07-02 00:00:00   bone   43
#> 87  242      8 1952 1952-08-01 12:00:00  blood   44
#> 88  242      8 1952 1952-08-01 12:00:00   bone   44
#> 89  209      9 1952 1952-09-01 00:00:01  blood   45
#> 90  209      9 1952 1952-09-01 00:00:01   bone   45
#> 91  191     10 1952 1952-10-01 12:00:00  blood   46
#> 92  191     10 1952 1952-10-01 12:00:00   bone   46
#> 93  172     11 1952 1952-11-01 00:00:00  blood   47
#> 94  172     11 1952 1952-11-01 00:00:00   bone   47
#> 95  194     12 1952 1952-12-01 12:00:01  blood   48
#> 96  194     12 1952 1952-12-01 12:00:01   bone   48
#> 97  196      1 1953 1953-01-01 00:00:00  blood   49
#> 98  196      1 1953 1953-01-01 00:00:00   bone   49
#> 99  196      2 1953 1953-01-31 10:00:00  blood   50
#> 100 196      2 1953 1953-01-31 10:00:00   bone   50
#> 101 236      3 1953 1953-03-02 20:00:01  blood   51
#> 102 236      3 1953 1953-03-02 20:00:01   bone   51
#> 103 235      4 1953 1953-04-02 06:00:00  blood   52
#> 104 235      4 1953 1953-04-02 06:00:00   bone   52
#> 105 229      5 1953 1953-05-02 16:00:00  blood   53
#> 106 229      5 1953 1953-05-02 16:00:00   bone   53
#> 107 243      6 1953 1953-06-02 02:00:01  blood   54
#> 108 243      6 1953 1953-06-02 02:00:01   bone   54
#> 109 264      7 1953 1953-07-02 12:00:00  blood   55
#> 110 264      7 1953 1953-07-02 12:00:00   bone   55
#> 111 272      8 1953 1953-08-01 22:00:00  blood   56
#> 112 272      8 1953 1953-08-01 22:00:00   bone   56
#> 113 237      9 1953 1953-09-01 08:00:01  blood   57
#> 114 237      9 1953 1953-09-01 08:00:01   bone   57
#> 115 211     10 1953 1953-10-01 18:00:00  blood   58
#> 116 211     10 1953 1953-10-01 18:00:00   bone   58
#> 117 180     11 1953 1953-11-01 04:00:00  blood   59
#> 118 180     11 1953 1953-11-01 04:00:00   bone   59
#> 119 201     12 1953 1953-12-01 14:00:01  blood   60
#> 120 201     12 1953 1953-12-01 14:00:01   bone   60
#> 121 204      1 1954 1954-01-01 00:00:00  blood   61
#> 122 204      1 1954 1954-01-01 00:00:00   bone   61
#> 123 188      2 1954 1954-01-31 10:00:00  blood   62
#> 124 188      2 1954 1954-01-31 10:00:00   bone   62
#> 125 235      3 1954 1954-03-02 20:00:01  blood   63
#> 126 235      3 1954 1954-03-02 20:00:01   bone   63
#> 127 227      4 1954 1954-04-02 06:00:00  blood   64
#> 128 227      4 1954 1954-04-02 06:00:00   bone   64
#> 129 234      5 1954 1954-05-02 16:00:00  blood   65
#> 130 234      5 1954 1954-05-02 16:00:00   bone   65
#> 131 264      6 1954 1954-06-02 02:00:01  blood   66
#> 132 264      6 1954 1954-06-02 02:00:01   bone   66
#> 133 302      7 1954 1954-07-02 12:00:00  blood   67
#> 134 302      7 1954 1954-07-02 12:00:00   bone   67
#> 135 293      8 1954 1954-08-01 22:00:00  blood   68
#> 136 293      8 1954 1954-08-01 22:00:00   bone   68
#> 137 259      9 1954 1954-09-01 08:00:01  blood   69
#> 138 259      9 1954 1954-09-01 08:00:01   bone   69
#> 139 229     10 1954 1954-10-01 18:00:00  blood   70
#> 140 229     10 1954 1954-10-01 18:00:00   bone   70
#> 141 203     11 1954 1954-11-01 04:00:01  blood   71
#> 142 203     11 1954 1954-11-01 04:00:01   bone   71
#> 143 229     12 1954 1954-12-01 14:00:01  blood   72
#> 144 229     12 1954 1954-12-01 14:00:01   bone   72
#> 145 242      1 1955 1955-01-01 00:00:00  blood   73
#> 146 242      1 1955 1955-01-01 00:00:00   bone   73
#> 147 233      2 1955 1955-01-31 10:00:01  blood   74
#> 148 233      2 1955 1955-01-31 10:00:01   bone   74
#> 149 267      3 1955 1955-03-02 20:00:01  blood   75
#> 150 267      3 1955 1955-03-02 20:00:01   bone   75
#> 151 269      4 1955 1955-04-02 06:00:00  blood   76
#> 152 269      4 1955 1955-04-02 06:00:00   bone   76
#> 153 270      5 1955 1955-05-02 16:00:01  blood   77
#> 154 270      5 1955 1955-05-02 16:00:01   bone   77
#> 155 315      6 1955 1955-06-02 02:00:01  blood   78
#> 156 315      6 1955 1955-06-02 02:00:01   bone   78
#> 157 364      7 1955 1955-07-02 12:00:00  blood   79
#> 158 364      7 1955 1955-07-02 12:00:00   bone   79
#> 159 347      8 1955 1955-08-01 22:00:01  blood   80
#> 160 347      8 1955 1955-08-01 22:00:01   bone   80
#> 161 312      9 1955 1955-09-01 08:00:01  blood   81
#> 162 312      9 1955 1955-09-01 08:00:01   bone   81
#> 163 274     10 1955 1955-10-01 18:00:00  blood   82
#> 164 274     10 1955 1955-10-01 18:00:00   bone   82
#> 165 237     11 1955 1955-11-01 04:00:01  blood   83
#> 166 237     11 1955 1955-11-01 04:00:01   bone   83
#> 167 278     12 1955 1955-12-01 14:00:01  blood   84
#> 168 278     12 1955 1955-12-01 14:00:01   bone   84
#> 169 284      1 1956 1956-01-01 00:00:00  blood   85
#> 170 284      1 1956 1956-01-01 00:00:00   bone   85
#> 171 277      2 1956 1956-01-31 12:00:01  blood   86
#> 172 277      2 1956 1956-01-31 12:00:01   bone   86
#> 173 317      3 1956 1956-03-02 00:00:01  blood   87
#> 174 317      3 1956 1956-03-02 00:00:01   bone   87
#> 175 313      4 1956 1956-04-01 12:00:00  blood   88
#> 176 313      4 1956 1956-04-01 12:00:00   bone   88
#> 177 318      5 1956 1956-05-02 00:00:01  blood   89
#> 178 318      5 1956 1956-05-02 00:00:01   bone   89
#> 179 374      6 1956 1956-06-01 12:00:01  blood   90
#> 180 374      6 1956 1956-06-01 12:00:01   bone   90
#> 181 413      7 1956 1956-07-02 00:00:00  blood   91
#> 182 413      7 1956 1956-07-02 00:00:00   bone   91
#> 183 405      8 1956 1956-08-01 12:00:01  blood   92
#> 184 405      8 1956 1956-08-01 12:00:01   bone   92
#> 185 355      9 1956 1956-09-01 00:00:01  blood   93
#> 186 355      9 1956 1956-09-01 00:00:01   bone   93
#> 187 306     10 1956 1956-10-01 12:00:00  blood   94
#> 188 306     10 1956 1956-10-01 12:00:00   bone   94
#> 189 271     11 1956 1956-11-01 00:00:01  blood   95
#> 190 271     11 1956 1956-11-01 00:00:01   bone   95
#> 191 306     12 1956 1956-12-01 12:00:01  blood   96
#> 192 306     12 1956 1956-12-01 12:00:01   bone   96
#> 193 315      1 1957 1957-01-01 00:00:00  blood   97
#> 194 315      1 1957 1957-01-01 00:00:00   bone   97
#> 195 301      2 1957 1957-01-31 10:00:01  blood   98
#> 196 301      2 1957 1957-01-31 10:00:01   bone   98
#> 197 356      3 1957 1957-03-02 20:00:01  blood   99
#> 198 356      3 1957 1957-03-02 20:00:01   bone   99
#> 199 348      4 1957 1957-04-02 06:00:00  blood  100
#> 200 348      4 1957 1957-04-02 06:00:00   bone  100
#> 201 355      5 1957 1957-05-02 16:00:01  blood  101
#> 202 355      5 1957 1957-05-02 16:00:01   bone  101
#> 203 422      6 1957 1957-06-02 02:00:01  blood  102
#> 204 422      6 1957 1957-06-02 02:00:01   bone  102
#> 205 465      7 1957 1957-07-02 12:00:00  blood  103
#> 206 465      7 1957 1957-07-02 12:00:00   bone  103
#> 207 467      8 1957 1957-08-01 22:00:01  blood  104
#> 208 467      8 1957 1957-08-01 22:00:01   bone  104
#> 209 404      9 1957 1957-09-01 08:00:01  blood  105
#> 210 404      9 1957 1957-09-01 08:00:01   bone  105
#> 211 347     10 1957 1957-10-01 18:00:00  blood  106
#> 212 347     10 1957 1957-10-01 18:00:00   bone  106
#> 213 305     11 1957 1957-11-01 04:00:01  blood  107
#> 214 305     11 1957 1957-11-01 04:00:01   bone  107
#> 215 336     12 1957 1957-12-01 14:00:01  blood  108
#> 216 336     12 1957 1957-12-01 14:00:01   bone  108
#> 217 340      1 1958 1958-01-01 00:00:00  blood  109
#> 218 340      1 1958 1958-01-01 00:00:00   bone  109
#> 219 318      2 1958 1958-01-31 10:00:01  blood  110
#> 220 318      2 1958 1958-01-31 10:00:01   bone  110
#> 221 362      3 1958 1958-03-02 20:00:01  blood  111
#> 222 362      3 1958 1958-03-02 20:00:01   bone  111
#> 223 348      4 1958 1958-04-02 06:00:00  blood  112
#> 224 348      4 1958 1958-04-02 06:00:00   bone  112
#> 225 363      5 1958 1958-05-02 16:00:01  blood  113
#> 226 363      5 1958 1958-05-02 16:00:01   bone  113
#> 227 435      6 1958 1958-06-02 02:00:01  blood  114
#> 228 435      6 1958 1958-06-02 02:00:01   bone  114
#> 229 491      7 1958 1958-07-02 12:00:00  blood  115
#> 230 491      7 1958 1958-07-02 12:00:00   bone  115
#> 231 505      8 1958 1958-08-01 22:00:01  blood  116
#> 232 505      8 1958 1958-08-01 22:00:01   bone  116
#> 233 404      9 1958 1958-09-01 08:00:01  blood  117
#> 234 404      9 1958 1958-09-01 08:00:01   bone  117
#> 235 359     10 1958 1958-10-01 18:00:00  blood  118
#> 236 359     10 1958 1958-10-01 18:00:00   bone  118
#> 237 310     11 1958 1958-11-01 04:00:01  blood  119
#> 238 310     11 1958 1958-11-01 04:00:01   bone  119
#> 239 337     12 1958 1958-12-01 14:00:01  blood  120
#> 240 337     12 1958 1958-12-01 14:00:01   bone  120
#> 241 360      1 1959 1959-01-01 00:00:00  blood  121
#> 242 360      1 1959 1959-01-01 00:00:00   bone  121
#> 243 342      2 1959 1959-01-31 10:00:01  blood  122
#> 244 342      2 1959 1959-01-31 10:00:01   bone  122
#> 
#> $data_test
#>      y season year                date series time
#> 1  406      3 1959 1959-03-02 20:00:01  blood  123
#> 2  406      3 1959 1959-03-02 20:00:01   bone  123
#> 3  396      4 1959 1959-04-02 06:00:00  blood  124
#> 4  396      4 1959 1959-04-02 06:00:00   bone  124
#> 5  420      5 1959 1959-05-02 16:00:01  blood  125
#> 6  420      5 1959 1959-05-02 16:00:01   bone  125
#> 7  472      6 1959 1959-06-02 02:00:01  blood  126
#> 8  472      6 1959 1959-06-02 02:00:01   bone  126
#> 9  548      7 1959 1959-07-02 12:00:00  blood  127
#> 10 548      7 1959 1959-07-02 12:00:00   bone  127
#> 11 559      8 1959 1959-08-01 22:00:01  blood  128
#> 12 559      8 1959 1959-08-01 22:00:01   bone  128
#> 13 463      9 1959 1959-09-01 08:00:01  blood  129
#> 14 463      9 1959 1959-09-01 08:00:01   bone  129
#> 15 407     10 1959 1959-10-01 18:00:00  blood  130
#> 16 407     10 1959 1959-10-01 18:00:00   bone  130
#> 17 362     11 1959 1959-11-01 04:00:01  blood  131
#> 18 362     11 1959 1959-11-01 04:00:01   bone  131
#> 19 405     12 1959 1959-12-01 14:00:01  blood  132
#> 20 405     12 1959 1959-12-01 14:00:01   bone  132
#> 21 417      1 1960 1960-01-01 00:00:00  blood  133
#> 22 417      1 1960 1960-01-01 00:00:00   bone  133
#> 23 391      2 1960 1960-01-31 12:00:01  blood  134
#> 24 391      2 1960 1960-01-31 12:00:01   bone  134
#> 25 419      3 1960 1960-03-02 00:00:01  blood  135
#> 26 419      3 1960 1960-03-02 00:00:01   bone  135
#> 27 461      4 1960 1960-04-01 12:00:00  blood  136
#> 28 461      4 1960 1960-04-01 12:00:00   bone  136
#> 29 472      5 1960 1960-05-02 00:00:01  blood  137
#> 30 472      5 1960 1960-05-02 00:00:01   bone  137
#> 31 535      6 1960 1960-06-01 12:00:01  blood  138
#> 32 535      6 1960 1960-06-01 12:00:01   bone  138
#> 33 622      7 1960 1960-07-02 00:00:00  blood  139
#> 34 622      7 1960 1960-07-02 00:00:00   bone  139
#> 35 606      8 1960 1960-08-01 12:00:01  blood  140
#> 36 606      8 1960 1960-08-01 12:00:01   bone  140
#> 37 508      9 1960 1960-09-01 00:00:01  blood  141
#> 38 508      9 1960 1960-09-01 00:00:01   bone  141
#> 39 461     10 1960 1960-10-01 12:00:00  blood  142
#> 40 461     10 1960 1960-10-01 12:00:00   bone  142
#> 41 390     11 1960 1960-11-01 00:00:01  blood  143
#> 42 390     11 1960 1960-11-01 00:00:01   bone  143
#> 43 432     12 1960 1960-12-01 12:00:01  blood  144
#> 44 432     12 1960 1960-12-01 12:00:01   bone  144
#> 

# An xts object example
dates <- seq(as.Date("2001-05-01"), length=30, by="quarter")
data  <- cbind(c(gas = rpois(30, cumprod(1+rnorm(30, mean = 0.01, sd = 0.001)))),
c(oil = rpois(30, cumprod(1+rnorm(30, mean = 0.01, sd = 0.001)))))
series <- xts(x = data, order.by = dates)
#> Error in xts(x = data, order.by = dates): could not find function "xts"
colnames(series) <- c('gas', 'oil')
head(series)
#>          gas oil
#> Jan 1949 112 112
#> Feb 1949 118 118
#> Mar 1949 132 132
#> Apr 1949 129 129
#> May 1949 121 121
#> Jun 1949 135 135
series_to_mvgam(series, freq = 4, train_prop = 0.85)
#> $data_train
#>       y season year                date series time
#> 1   112      1 1949 1949-01-01 00:00:00    gas    1
#> 2   112      1 1949 1949-01-01 00:00:00    oil    1
#> 3   118      2 1949 1949-01-31 10:00:00    gas    2
#> 4   118      2 1949 1949-01-31 10:00:00    oil    2
#> 5   132      3 1949 1949-03-02 20:00:01    gas    3
#> 6   132      3 1949 1949-03-02 20:00:01    oil    3
#> 7   129      4 1949 1949-04-02 06:00:00    gas    4
#> 8   129      4 1949 1949-04-02 06:00:00    oil    4
#> 9   121      5 1949 1949-05-02 16:00:00    gas    5
#> 10  121      5 1949 1949-05-02 16:00:00    oil    5
#> 11  135      6 1949 1949-06-02 02:00:01    gas    6
#> 12  135      6 1949 1949-06-02 02:00:01    oil    6
#> 13  148      7 1949 1949-07-02 12:00:00    gas    7
#> 14  148      7 1949 1949-07-02 12:00:00    oil    7
#> 15  148      8 1949 1949-08-01 22:00:00    gas    8
#> 16  148      8 1949 1949-08-01 22:00:00    oil    8
#> 17  136      9 1949 1949-09-01 08:00:01    gas    9
#> 18  136      9 1949 1949-09-01 08:00:01    oil    9
#> 19  119     10 1949 1949-10-01 18:00:00    gas   10
#> 20  119     10 1949 1949-10-01 18:00:00    oil   10
#> 21  104     11 1949 1949-11-01 04:00:00    gas   11
#> 22  104     11 1949 1949-11-01 04:00:00    oil   11
#> 23  118     12 1949 1949-12-01 14:00:01    gas   12
#> 24  118     12 1949 1949-12-01 14:00:01    oil   12
#> 25  115      1 1950 1950-01-01 00:00:00    gas   13
#> 26  115      1 1950 1950-01-01 00:00:00    oil   13
#> 27  126      2 1950 1950-01-31 10:00:00    gas   14
#> 28  126      2 1950 1950-01-31 10:00:00    oil   14
#> 29  141      3 1950 1950-03-02 20:00:01    gas   15
#> 30  141      3 1950 1950-03-02 20:00:01    oil   15
#> 31  135      4 1950 1950-04-02 06:00:00    gas   16
#> 32  135      4 1950 1950-04-02 06:00:00    oil   16
#> 33  125      5 1950 1950-05-02 16:00:00    gas   17
#> 34  125      5 1950 1950-05-02 16:00:00    oil   17
#> 35  149      6 1950 1950-06-02 02:00:01    gas   18
#> 36  149      6 1950 1950-06-02 02:00:01    oil   18
#> 37  170      7 1950 1950-07-02 12:00:00    gas   19
#> 38  170      7 1950 1950-07-02 12:00:00    oil   19
#> 39  170      8 1950 1950-08-01 22:00:00    gas   20
#> 40  170      8 1950 1950-08-01 22:00:00    oil   20
#> 41  158      9 1950 1950-09-01 08:00:01    gas   21
#> 42  158      9 1950 1950-09-01 08:00:01    oil   21
#> 43  133     10 1950 1950-10-01 18:00:00    gas   22
#> 44  133     10 1950 1950-10-01 18:00:00    oil   22
#> 45  114     11 1950 1950-11-01 04:00:00    gas   23
#> 46  114     11 1950 1950-11-01 04:00:00    oil   23
#> 47  140     12 1950 1950-12-01 14:00:01    gas   24
#> 48  140     12 1950 1950-12-01 14:00:01    oil   24
#> 49  145      1 1951 1951-01-01 00:00:00    gas   25
#> 50  145      1 1951 1951-01-01 00:00:00    oil   25
#> 51  150      2 1951 1951-01-31 10:00:00    gas   26
#> 52  150      2 1951 1951-01-31 10:00:00    oil   26
#> 53  178      3 1951 1951-03-02 20:00:01    gas   27
#> 54  178      3 1951 1951-03-02 20:00:01    oil   27
#> 55  163      4 1951 1951-04-02 06:00:00    gas   28
#> 56  163      4 1951 1951-04-02 06:00:00    oil   28
#> 57  172      5 1951 1951-05-02 16:00:00    gas   29
#> 58  172      5 1951 1951-05-02 16:00:00    oil   29
#> 59  178      6 1951 1951-06-02 02:00:01    gas   30
#> 60  178      6 1951 1951-06-02 02:00:01    oil   30
#> 61  199      7 1951 1951-07-02 12:00:00    gas   31
#> 62  199      7 1951 1951-07-02 12:00:00    oil   31
#> 63  199      8 1951 1951-08-01 22:00:00    gas   32
#> 64  199      8 1951 1951-08-01 22:00:00    oil   32
#> 65  184      9 1951 1951-09-01 08:00:01    gas   33
#> 66  184      9 1951 1951-09-01 08:00:01    oil   33
#> 67  162     10 1951 1951-10-01 18:00:00    gas   34
#> 68  162     10 1951 1951-10-01 18:00:00    oil   34
#> 69  146     11 1951 1951-11-01 04:00:00    gas   35
#> 70  146     11 1951 1951-11-01 04:00:00    oil   35
#> 71  166     12 1951 1951-12-01 14:00:01    gas   36
#> 72  166     12 1951 1951-12-01 14:00:01    oil   36
#> 73  171      1 1952 1952-01-01 00:00:00    gas   37
#> 74  171      1 1952 1952-01-01 00:00:00    oil   37
#> 75  180      2 1952 1952-01-31 12:00:00    gas   38
#> 76  180      2 1952 1952-01-31 12:00:00    oil   38
#> 77  193      3 1952 1952-03-02 00:00:01    gas   39
#> 78  193      3 1952 1952-03-02 00:00:01    oil   39
#> 79  181      4 1952 1952-04-01 12:00:00    gas   40
#> 80  181      4 1952 1952-04-01 12:00:00    oil   40
#> 81  183      5 1952 1952-05-02 00:00:00    gas   41
#> 82  183      5 1952 1952-05-02 00:00:00    oil   41
#> 83  218      6 1952 1952-06-01 12:00:01    gas   42
#> 84  218      6 1952 1952-06-01 12:00:01    oil   42
#> 85  230      7 1952 1952-07-02 00:00:00    gas   43
#> 86  230      7 1952 1952-07-02 00:00:00    oil   43
#> 87  242      8 1952 1952-08-01 12:00:00    gas   44
#> 88  242      8 1952 1952-08-01 12:00:00    oil   44
#> 89  209      9 1952 1952-09-01 00:00:01    gas   45
#> 90  209      9 1952 1952-09-01 00:00:01    oil   45
#> 91  191     10 1952 1952-10-01 12:00:00    gas   46
#> 92  191     10 1952 1952-10-01 12:00:00    oil   46
#> 93  172     11 1952 1952-11-01 00:00:00    gas   47
#> 94  172     11 1952 1952-11-01 00:00:00    oil   47
#> 95  194     12 1952 1952-12-01 12:00:01    gas   48
#> 96  194     12 1952 1952-12-01 12:00:01    oil   48
#> 97  196      1 1953 1953-01-01 00:00:00    gas   49
#> 98  196      1 1953 1953-01-01 00:00:00    oil   49
#> 99  196      2 1953 1953-01-31 10:00:00    gas   50
#> 100 196      2 1953 1953-01-31 10:00:00    oil   50
#> 101 236      3 1953 1953-03-02 20:00:01    gas   51
#> 102 236      3 1953 1953-03-02 20:00:01    oil   51
#> 103 235      4 1953 1953-04-02 06:00:00    gas   52
#> 104 235      4 1953 1953-04-02 06:00:00    oil   52
#> 105 229      5 1953 1953-05-02 16:00:00    gas   53
#> 106 229      5 1953 1953-05-02 16:00:00    oil   53
#> 107 243      6 1953 1953-06-02 02:00:01    gas   54
#> 108 243      6 1953 1953-06-02 02:00:01    oil   54
#> 109 264      7 1953 1953-07-02 12:00:00    gas   55
#> 110 264      7 1953 1953-07-02 12:00:00    oil   55
#> 111 272      8 1953 1953-08-01 22:00:00    gas   56
#> 112 272      8 1953 1953-08-01 22:00:00    oil   56
#> 113 237      9 1953 1953-09-01 08:00:01    gas   57
#> 114 237      9 1953 1953-09-01 08:00:01    oil   57
#> 115 211     10 1953 1953-10-01 18:00:00    gas   58
#> 116 211     10 1953 1953-10-01 18:00:00    oil   58
#> 117 180     11 1953 1953-11-01 04:00:00    gas   59
#> 118 180     11 1953 1953-11-01 04:00:00    oil   59
#> 119 201     12 1953 1953-12-01 14:00:01    gas   60
#> 120 201     12 1953 1953-12-01 14:00:01    oil   60
#> 121 204      1 1954 1954-01-01 00:00:00    gas   61
#> 122 204      1 1954 1954-01-01 00:00:00    oil   61
#> 123 188      2 1954 1954-01-31 10:00:00    gas   62
#> 124 188      2 1954 1954-01-31 10:00:00    oil   62
#> 125 235      3 1954 1954-03-02 20:00:01    gas   63
#> 126 235      3 1954 1954-03-02 20:00:01    oil   63
#> 127 227      4 1954 1954-04-02 06:00:00    gas   64
#> 128 227      4 1954 1954-04-02 06:00:00    oil   64
#> 129 234      5 1954 1954-05-02 16:00:00    gas   65
#> 130 234      5 1954 1954-05-02 16:00:00    oil   65
#> 131 264      6 1954 1954-06-02 02:00:01    gas   66
#> 132 264      6 1954 1954-06-02 02:00:01    oil   66
#> 133 302      7 1954 1954-07-02 12:00:00    gas   67
#> 134 302      7 1954 1954-07-02 12:00:00    oil   67
#> 135 293      8 1954 1954-08-01 22:00:00    gas   68
#> 136 293      8 1954 1954-08-01 22:00:00    oil   68
#> 137 259      9 1954 1954-09-01 08:00:01    gas   69
#> 138 259      9 1954 1954-09-01 08:00:01    oil   69
#> 139 229     10 1954 1954-10-01 18:00:00    gas   70
#> 140 229     10 1954 1954-10-01 18:00:00    oil   70
#> 141 203     11 1954 1954-11-01 04:00:01    gas   71
#> 142 203     11 1954 1954-11-01 04:00:01    oil   71
#> 143 229     12 1954 1954-12-01 14:00:01    gas   72
#> 144 229     12 1954 1954-12-01 14:00:01    oil   72
#> 145 242      1 1955 1955-01-01 00:00:00    gas   73
#> 146 242      1 1955 1955-01-01 00:00:00    oil   73
#> 147 233      2 1955 1955-01-31 10:00:01    gas   74
#> 148 233      2 1955 1955-01-31 10:00:01    oil   74
#> 149 267      3 1955 1955-03-02 20:00:01    gas   75
#> 150 267      3 1955 1955-03-02 20:00:01    oil   75
#> 151 269      4 1955 1955-04-02 06:00:00    gas   76
#> 152 269      4 1955 1955-04-02 06:00:00    oil   76
#> 153 270      5 1955 1955-05-02 16:00:01    gas   77
#> 154 270      5 1955 1955-05-02 16:00:01    oil   77
#> 155 315      6 1955 1955-06-02 02:00:01    gas   78
#> 156 315      6 1955 1955-06-02 02:00:01    oil   78
#> 157 364      7 1955 1955-07-02 12:00:00    gas   79
#> 158 364      7 1955 1955-07-02 12:00:00    oil   79
#> 159 347      8 1955 1955-08-01 22:00:01    gas   80
#> 160 347      8 1955 1955-08-01 22:00:01    oil   80
#> 161 312      9 1955 1955-09-01 08:00:01    gas   81
#> 162 312      9 1955 1955-09-01 08:00:01    oil   81
#> 163 274     10 1955 1955-10-01 18:00:00    gas   82
#> 164 274     10 1955 1955-10-01 18:00:00    oil   82
#> 165 237     11 1955 1955-11-01 04:00:01    gas   83
#> 166 237     11 1955 1955-11-01 04:00:01    oil   83
#> 167 278     12 1955 1955-12-01 14:00:01    gas   84
#> 168 278     12 1955 1955-12-01 14:00:01    oil   84
#> 169 284      1 1956 1956-01-01 00:00:00    gas   85
#> 170 284      1 1956 1956-01-01 00:00:00    oil   85
#> 171 277      2 1956 1956-01-31 12:00:01    gas   86
#> 172 277      2 1956 1956-01-31 12:00:01    oil   86
#> 173 317      3 1956 1956-03-02 00:00:01    gas   87
#> 174 317      3 1956 1956-03-02 00:00:01    oil   87
#> 175 313      4 1956 1956-04-01 12:00:00    gas   88
#> 176 313      4 1956 1956-04-01 12:00:00    oil   88
#> 177 318      5 1956 1956-05-02 00:00:01    gas   89
#> 178 318      5 1956 1956-05-02 00:00:01    oil   89
#> 179 374      6 1956 1956-06-01 12:00:01    gas   90
#> 180 374      6 1956 1956-06-01 12:00:01    oil   90
#> 181 413      7 1956 1956-07-02 00:00:00    gas   91
#> 182 413      7 1956 1956-07-02 00:00:00    oil   91
#> 183 405      8 1956 1956-08-01 12:00:01    gas   92
#> 184 405      8 1956 1956-08-01 12:00:01    oil   92
#> 185 355      9 1956 1956-09-01 00:00:01    gas   93
#> 186 355      9 1956 1956-09-01 00:00:01    oil   93
#> 187 306     10 1956 1956-10-01 12:00:00    gas   94
#> 188 306     10 1956 1956-10-01 12:00:00    oil   94
#> 189 271     11 1956 1956-11-01 00:00:01    gas   95
#> 190 271     11 1956 1956-11-01 00:00:01    oil   95
#> 191 306     12 1956 1956-12-01 12:00:01    gas   96
#> 192 306     12 1956 1956-12-01 12:00:01    oil   96
#> 193 315      1 1957 1957-01-01 00:00:00    gas   97
#> 194 315      1 1957 1957-01-01 00:00:00    oil   97
#> 195 301      2 1957 1957-01-31 10:00:01    gas   98
#> 196 301      2 1957 1957-01-31 10:00:01    oil   98
#> 197 356      3 1957 1957-03-02 20:00:01    gas   99
#> 198 356      3 1957 1957-03-02 20:00:01    oil   99
#> 199 348      4 1957 1957-04-02 06:00:00    gas  100
#> 200 348      4 1957 1957-04-02 06:00:00    oil  100
#> 201 355      5 1957 1957-05-02 16:00:01    gas  101
#> 202 355      5 1957 1957-05-02 16:00:01    oil  101
#> 203 422      6 1957 1957-06-02 02:00:01    gas  102
#> 204 422      6 1957 1957-06-02 02:00:01    oil  102
#> 205 465      7 1957 1957-07-02 12:00:00    gas  103
#> 206 465      7 1957 1957-07-02 12:00:00    oil  103
#> 207 467      8 1957 1957-08-01 22:00:01    gas  104
#> 208 467      8 1957 1957-08-01 22:00:01    oil  104
#> 209 404      9 1957 1957-09-01 08:00:01    gas  105
#> 210 404      9 1957 1957-09-01 08:00:01    oil  105
#> 211 347     10 1957 1957-10-01 18:00:00    gas  106
#> 212 347     10 1957 1957-10-01 18:00:00    oil  106
#> 213 305     11 1957 1957-11-01 04:00:01    gas  107
#> 214 305     11 1957 1957-11-01 04:00:01    oil  107
#> 215 336     12 1957 1957-12-01 14:00:01    gas  108
#> 216 336     12 1957 1957-12-01 14:00:01    oil  108
#> 217 340      1 1958 1958-01-01 00:00:00    gas  109
#> 218 340      1 1958 1958-01-01 00:00:00    oil  109
#> 219 318      2 1958 1958-01-31 10:00:01    gas  110
#> 220 318      2 1958 1958-01-31 10:00:01    oil  110
#> 221 362      3 1958 1958-03-02 20:00:01    gas  111
#> 222 362      3 1958 1958-03-02 20:00:01    oil  111
#> 223 348      4 1958 1958-04-02 06:00:00    gas  112
#> 224 348      4 1958 1958-04-02 06:00:00    oil  112
#> 225 363      5 1958 1958-05-02 16:00:01    gas  113
#> 226 363      5 1958 1958-05-02 16:00:01    oil  113
#> 227 435      6 1958 1958-06-02 02:00:01    gas  114
#> 228 435      6 1958 1958-06-02 02:00:01    oil  114
#> 229 491      7 1958 1958-07-02 12:00:00    gas  115
#> 230 491      7 1958 1958-07-02 12:00:00    oil  115
#> 231 505      8 1958 1958-08-01 22:00:01    gas  116
#> 232 505      8 1958 1958-08-01 22:00:01    oil  116
#> 233 404      9 1958 1958-09-01 08:00:01    gas  117
#> 234 404      9 1958 1958-09-01 08:00:01    oil  117
#> 235 359     10 1958 1958-10-01 18:00:00    gas  118
#> 236 359     10 1958 1958-10-01 18:00:00    oil  118
#> 237 310     11 1958 1958-11-01 04:00:01    gas  119
#> 238 310     11 1958 1958-11-01 04:00:01    oil  119
#> 239 337     12 1958 1958-12-01 14:00:01    gas  120
#> 240 337     12 1958 1958-12-01 14:00:01    oil  120
#> 241 360      1 1959 1959-01-01 00:00:00    gas  121
#> 242 360      1 1959 1959-01-01 00:00:00    oil  121
#> 243 342      2 1959 1959-01-31 10:00:01    gas  122
#> 244 342      2 1959 1959-01-31 10:00:01    oil  122
#> 
#> $data_test
#>      y season year                date series time
#> 1  406      3 1959 1959-03-02 20:00:01    gas  123
#> 2  406      3 1959 1959-03-02 20:00:01    oil  123
#> 3  396      4 1959 1959-04-02 06:00:00    gas  124
#> 4  396      4 1959 1959-04-02 06:00:00    oil  124
#> 5  420      5 1959 1959-05-02 16:00:01    gas  125
#> 6  420      5 1959 1959-05-02 16:00:01    oil  125
#> 7  472      6 1959 1959-06-02 02:00:01    gas  126
#> 8  472      6 1959 1959-06-02 02:00:01    oil  126
#> 9  548      7 1959 1959-07-02 12:00:00    gas  127
#> 10 548      7 1959 1959-07-02 12:00:00    oil  127
#> 11 559      8 1959 1959-08-01 22:00:01    gas  128
#> 12 559      8 1959 1959-08-01 22:00:01    oil  128
#> 13 463      9 1959 1959-09-01 08:00:01    gas  129
#> 14 463      9 1959 1959-09-01 08:00:01    oil  129
#> 15 407     10 1959 1959-10-01 18:00:00    gas  130
#> 16 407     10 1959 1959-10-01 18:00:00    oil  130
#> 17 362     11 1959 1959-11-01 04:00:01    gas  131
#> 18 362     11 1959 1959-11-01 04:00:01    oil  131
#> 19 405     12 1959 1959-12-01 14:00:01    gas  132
#> 20 405     12 1959 1959-12-01 14:00:01    oil  132
#> 21 417      1 1960 1960-01-01 00:00:00    gas  133
#> 22 417      1 1960 1960-01-01 00:00:00    oil  133
#> 23 391      2 1960 1960-01-31 12:00:01    gas  134
#> 24 391      2 1960 1960-01-31 12:00:01    oil  134
#> 25 419      3 1960 1960-03-02 00:00:01    gas  135
#> 26 419      3 1960 1960-03-02 00:00:01    oil  135
#> 27 461      4 1960 1960-04-01 12:00:00    gas  136
#> 28 461      4 1960 1960-04-01 12:00:00    oil  136
#> 29 472      5 1960 1960-05-02 00:00:01    gas  137
#> 30 472      5 1960 1960-05-02 00:00:01    oil  137
#> 31 535      6 1960 1960-06-01 12:00:01    gas  138
#> 32 535      6 1960 1960-06-01 12:00:01    oil  138
#> 33 622      7 1960 1960-07-02 00:00:00    gas  139
#> 34 622      7 1960 1960-07-02 00:00:00    oil  139
#> 35 606      8 1960 1960-08-01 12:00:01    gas  140
#> 36 606      8 1960 1960-08-01 12:00:01    oil  140
#> 37 508      9 1960 1960-09-01 00:00:01    gas  141
#> 38 508      9 1960 1960-09-01 00:00:01    oil  141
#> 39 461     10 1960 1960-10-01 12:00:00    gas  142
#> 40 461     10 1960 1960-10-01 12:00:00    oil  142
#> 41 390     11 1960 1960-11-01 00:00:01    gas  143
#> 42 390     11 1960 1960-11-01 00:00:01    oil  143
#> 43 432     12 1960 1960-12-01 12:00:01    gas  144
#> 44 432     12 1960 1960-12-01 12:00:01    oil  144
#>