Generate a methods description for mvgam models

Create a brief but fully referenced methods description, along with a useful list of references, for fitted mvgam and jsdgam models.

Usage

how_to_cite(object, ...)

# S3 method for mvgam
how_to_cite(object, ...)

Arguments

object

list object of class mvgam resulting from a call to mvgam() or jsdgam()

...

ignored

Value

An object of class how_to_cite containing a text description of the methods as well as lists of both primary and additional references.

Details

This function uses the model's structure to come up with a very basic but hopefully useful methods description that can help users to appropriately acknowledge the hard work of developers and champion open science. Please do not consider the text returned by this function to be a completely adequate methods section; it is only meant to get you started.

See also

Author

Nicholas J Clark

Examples

# \dontrun{
#--------------------------------------------------
# Simulate 4 time series with hierarchical seasonality
# and a VAR(1) dynamic process
#--------------------------------------------------
set.seed(0)

simdat <- sim_mvgam(
  seasonality = 'hierarchical',
  trend_model = VAR(cor = TRUE),
  family = gaussian()
)

# Fit an appropriate model
mod1 <- mvgam(
  y ~ s(season, bs = 'cc', k = 6),
  data = simdat$data_train,
  family = gaussian(),
  trend_model = VAR(cor = TRUE),
  chains = 2,
  silent = 2
)

how_to_cite(mod1)
#> Methods text skeleton
#> We used the R package mvgam (version 1.1.595; Clark & Wells, 2023) to
#>   construct, fit and interrogate the model. mvgam fits Bayesian
#>   State-Space models that can include flexible predictor effects in both
#>   the process and observation components by incorporating functionalities
#>   from the brms (Burkner 2017), mgcv (Wood 2017) and splines2 (Wang & Yan,
#>   2023) packages. To encourage stability and prevent forecast variance
#>   from increasing indefinitely, we enforced stationarity of the Vector
#>   Autoregressive process following methods described by Heaps (2023) and
#>   Clark et al. (2025). The mvgam-constructed model and observed data were
#>   passed to the probabilistic programming environment Stan (version
#>   2.38.0; Carpenter et al. 2017, Stan Development Team 2026), specifically
#>   through the cmdstanr interface (Gabry & Cesnovar, 2021). We ran 2
#>   Hamiltonian Monte Carlo chains for 500 warmup iterations and 500
#>   sampling iterations for joint posterior estimation. Rank normalized
#>   split Rhat (Vehtari et al. 2021) and effective sample sizes were used to
#>   monitor convergence.
#> 
#> Primary references
#> Clark, NJ and Wells K (2023). Dynamic Generalized Additive Models
#>   (DGAMs) for forecasting discrete ecological time series. Methods in
#>   Ecology and Evolution, 14, 771-784. doi.org/10.1111/2041-210X.13974
#> Burkner, PC (2017). brms: An R Package for Bayesian Multilevel Models
#>   Using Stan. Journal of Statistical Software, 80(1), 1-28.
#>   doi:10.18637/jss.v080.i01
#> Wood, SN (2017). Generalized Additive Models: An Introduction with R
#>   (2nd edition). Chapman and Hall/CRC.
#> Wang W and Yan J (2021). Shape-Restricted Regression Splines with R
#>   Package splines2. Journal of Data Science, 19(3), 498-517.
#>   doi:10.6339/21-JDS1020 https://doi.org/10.6339/21-JDS1020.
#> Heaps, SE (2023). Enforcing stationarity through the prior in vector
#>   autoregressions. Journal of Computational and Graphical Statistics 32,
#>   74-83.
#> Clark NJ, Ernest SKM, Senyondo H, Simonis J, White EP, Yenni GM,
#>   Karunarathna KANK (2025). Beyond single-species models: leveraging
#>   multispecies forecasts to navigate the dynamics of ecological
#>   predictability. PeerJ 13:e18929.
#> Carpenter B, Gelman A, Hoffman MD, Lee D, Goodrich B, Betancourt M,
#>   Brubaker M, Guo J, Li P and Riddell A (2017). Stan: A probabilistic
#>   programming language. Journal of Statistical Software 76.
#> Gabry J, Cesnovar R, Johnson A, and Bronder S (2026). cmdstanr: R
#>   Interface to 'CmdStan'. https://mc-stan.org/cmdstanr/,
#>   https://discourse.mc-stan.org.
#> Vehtari A, Gelman A, Simpson D, Carpenter B, and Burkner P (2021).
#>   Rank-normalization, folding, and localization: An improved Rhat for
#>   assessing convergence of MCMC (with discussion). Bayesian Analysis 16(2)
#>   667-718. https://doi.org/10.1214/20-BA1221.
#> 
#> Other useful references
#> Arel-Bundock V, Greifer N, and Heiss A (2024). How to interpret
#>   statistical models using marginaleffects for R and Python. Journal of
#>   Statistical Software, 111(9), 1-32.
#>   https://doi.org/10.18637/jss.v111.i09
#> Gabry J, Simpson D, Vehtari A, Betancourt M, and Gelman A (2019).
#>   Visualization in Bayesian workflow. Journal of the Royal Statatistical
#>   Society A, 182, 389-402. doi:10.1111/rssa.12378.
#> Vehtari A, Gelman A, and Gabry J (2017). Practical Bayesian model
#>   evaluation using leave-one-out cross-validation and WAIC. Statistics and
#>   Computing, 27, 1413-1432. doi:10.1007/s11222-016-9696-4.
#> Burkner PC, Gabry J, and Vehtari A. (2020). Approximate leave-future-out
#>   cross-validation for Bayesian time series models. Journal of Statistical
#>   Computation and Simulation, 90(14), 2499-2523.
#>   https://doi.org/10.1080/00949655.2020.1783262

#--------------------------------------------------
# For a GP example, simulate data using the mgcv package
#--------------------------------------------------
dat <- mgcv::gamSim(1, n = 30, scale = 2)
#> Gu & Wahba 4 term additive model

# Fit a model that uses an approximate GP from brms
mod2 <- mvgam(
  y ~ gp(x2, k = 12),
  data = dat,
  family = gaussian(),
  chains = 2,
  silent = 2
)
#> Warning: gp effects in mvgam cannot yet handle autogrouping
#> resetting all instances of 'gr = TRUE' to 'gr = FALSE'
#> This warning is displayed once per session.

how_to_cite(mod2)
#> Methods text skeleton
#> We used the R package mvgam (version 1.1.595; Clark & Wells, 2023) to
#>   construct, fit and interrogate the model. mvgam fits Bayesian
#>   State-Space models that can include flexible predictor effects in both
#>   the process and observation components by incorporating functionalities
#>   from the brms (Burkner 2017), mgcv (Wood 2017) and splines2 (Wang & Yan,
#>   2023) packages. Gaussian Process functional effects were estimated using
#>   a low-rank Hilbert space approximation following methods described by
#>   Riutort-Mayol et al. (2023). The mvgam-constructed model and observed
#>   data were passed to the probabilistic programming environment Stan
#>   (version 2.38.0; Carpenter et al. 2017, Stan Development Team 2026),
#>   specifically through the cmdstanr interface (Gabry & Cesnovar, 2021). We
#>   ran 2 Hamiltonian Monte Carlo chains for 500 warmup iterations and 500
#>   sampling iterations for joint posterior estimation. Rank normalized
#>   split Rhat (Vehtari et al. 2021) and effective sample sizes were used to
#>   monitor convergence.
#> 
#> Primary references
#> Clark, NJ and Wells K (2023). Dynamic Generalized Additive Models
#>   (DGAMs) for forecasting discrete ecological time series. Methods in
#>   Ecology and Evolution, 14, 771-784. doi.org/10.1111/2041-210X.13974
#> Burkner, PC (2017). brms: An R Package for Bayesian Multilevel Models
#>   Using Stan. Journal of Statistical Software, 80(1), 1-28.
#>   doi:10.18637/jss.v080.i01
#> Wood, SN (2017). Generalized Additive Models: An Introduction with R
#>   (2nd edition). Chapman and Hall/CRC.
#> Wang W and Yan J (2021). Shape-Restricted Regression Splines with R
#>   Package splines2. Journal of Data Science, 19(3), 498-517.
#>   doi:10.6339/21-JDS1020 https://doi.org/10.6339/21-JDS1020.
#> Riutort-Mayol G, Burkner PC, Andersen MR, Solin A and Vehtari A (2023).
#>   Practical Hilbert space approximate Bayesian Gaussian processes for
#>   probabilistic programming. Statistics and Computing 33, 1.
#>   https://doi.org/10.1007/s11222-022-10167-2
#> Carpenter B, Gelman A, Hoffman MD, Lee D, Goodrich B, Betancourt M,
#>   Brubaker M, Guo J, Li P and Riddell A (2017). Stan: A probabilistic
#>   programming language. Journal of Statistical Software 76.
#> Gabry J, Cesnovar R, Johnson A, and Bronder S (2026). cmdstanr: R
#>   Interface to 'CmdStan'. https://mc-stan.org/cmdstanr/,
#>   https://discourse.mc-stan.org.
#> Vehtari A, Gelman A, Simpson D, Carpenter B, and Burkner P (2021).
#>   Rank-normalization, folding, and localization: An improved Rhat for
#>   assessing convergence of MCMC (with discussion). Bayesian Analysis 16(2)
#>   667-718. https://doi.org/10.1214/20-BA1221.
#> 
#> Other useful references
#> Arel-Bundock V, Greifer N, and Heiss A (2024). How to interpret
#>   statistical models using marginaleffects for R and Python. Journal of
#>   Statistical Software, 111(9), 1-32.
#>   https://doi.org/10.18637/jss.v111.i09
#> Gabry J, Simpson D, Vehtari A, Betancourt M, and Gelman A (2019).
#>   Visualization in Bayesian workflow. Journal of the Royal Statatistical
#>   Society A, 182, 389-402. doi:10.1111/rssa.12378.
#> Vehtari A, Gelman A, and Gabry J (2017). Practical Bayesian model
#>   evaluation using leave-one-out cross-validation and WAIC. Statistics and
#>   Computing, 27, 1413-1432. doi:10.1007/s11222-016-9696-4.
#> Burkner PC, Gabry J, and Vehtari A. (2020). Approximate leave-future-out
#>   cross-validation for Bayesian time series models. Journal of Statistical
#>   Computation and Simulation, 90(14), 2499-2523.
#>   https://doi.org/10.1080/00949655.2020.1783262
# }