Posterior draws of residuals from mvgam models

This method extracts posterior draws of Dunn-Smyth (randomized quantile) residuals in the order in which the data were supplied to the model. It included additional arguments for obtaining summaries of the computed residuals

Usage

# S3 method for class 'mvgam'
residuals(object, summary = TRUE, robust = FALSE, probs = c(0.025, 0.975), ...)

Arguments

object

An object of class mvgam

summary

Should summary statistics be returned instead of the raw values? Default is TRUE..

robust

If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs

The percentiles to be computed by the quantile function. Only used if summary is TRUE.

...

ignored

Value

An array of randomized quantile residual values. If summary = FALSE the output resembles those of posterior_epred.mvgam and predict.mvgam.

If summary = TRUE the output is an n_observations x E matrix. The number of summary statistics E is equal to 2 + length(probs): The Estimate column contains point estimates (either mean or median depending on argument robust), while the Est.Error column contains uncertainty estimates (either standard deviation or median absolute deviation depending on argument robust). The remaining columns starting with Q contain quantile estimates as specified via argument probs.

Details

This method gives residuals as Dunn-Smyth (randomized quantile) residuals. Any observations that were missing (i.e. NA) in the original data will have missing values in the residuals

See also

Author

Nicholas J Clark

Examples

# \donttest{
# Simulate some data and fit a model
simdat <- sim_mvgam(n_series = 1, trend_model = AR())
mod <- mvgam(y ~ s(season, bs = 'cc'),
             trend_model = AR(),
             noncentred = TRUE,
             data = simdat$data_train,
             chains = 2,
             silent = 2)

# Extract posterior residuals
resids <- residuals(mod)
str(resids)
#>  num [1:75, 1:4] 0.7284 0.228 0.7504 -0.579 -0.0936 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : NULL
#>   ..$ : chr [1:4] "Estimate" "Est.Error" "Q2.5" "Q97.5"

# Or add them directly to the observed data, along with fitted values
augment(mod, robust = FALSE, probs = c(0.25, 0.75))
#> # A tibble: 75 × 14
#>        y season  year series    time .observed .fitted .fit.variability .fit.cred.low .fit.cred.high  .resid
#>    <int>  <int> <int> <fct>    <int>     <int>   <dbl>            <dbl>         <dbl>          <dbl>   <dbl>
#>  1     1      1     1 series_1     1         1   0.542            0.257         0.373          0.645  0.728 
#>  2     1      2     1 series_1     2         1   0.962            0.382         0.703          1.16   0.228 
#>  3     3      3     1 series_1     3         3   2.06             0.745         1.58           2.42   0.750 
#>  4     1      4     1 series_1     4         1   1.97             0.703         1.49           2.40  -0.579 
#>  5     1      5     1 series_1     5         1   1.32             0.497         0.965          1.62  -0.0936
#>  6     0      6     1 series_1     6         0   0.673            0.298         0.464          0.835 -0.736 
#>  7     0      7     1 series_1     7         0   0.510            0.240         0.346          0.620 -0.645 
#>  8     1      8     1 series_1     8         1   0.518            0.265         0.332          0.636  0.776 
#>  9     0      9     1 series_1     9         0   0.418            0.204         0.279          0.514 -0.583 
#> 10     0     10     1 series_1    10         0   0.381            0.187         0.253          0.487 -0.482 
#> # ℹ 65 more rows
#> # ℹ 3 more variables: .resid.variability <dbl>, .resid.cred.low <dbl>, .resid.cred.high <dbl>
# }