Summary for a fitted mvgam object

These functions take a fitted mvgam object and return various useful summaries

Usage

# S3 method for mvgam
summary(object, include_betas = TRUE, digits = 2, ...)

# S3 method for mvgam_prefit
summary(object, ...)

# S3 method for mvgam
coef(object, summarise = TRUE, ...)

Arguments

object

list object returned from mvgam

include_betas

Logical. Print a summary of linear predictor beta coefficients? Defaults to TRUE, but if your model contains many spline coefficients then it may be worth setting this to FALSE for a more concise summary

digits

The number of significant digits for printing out the summary; defaults to 2.

...

Ignored

summarise

logical. Summaries of coefficients will be returned if TRUE. Otherwise the full posterior distribution will be returned

Value

For summary.mvgam and summary.mvgam_prefit, Aalist is printed on-screen showing the summaries for the model

For coef.mvgam, either a matrix of posterior coefficient distributions (if summarise == FALSE or data.frame of coefficient summaries)

Details

summary.mvgam and summary.mvgam_prefit return brief summaries of the model's call is printed, along with posterior intervals for some of the key parameters in the model. Note that some smooths have extra penalties on the null space, so summaries for the rho parameters may include more penalty terms than the number of smooths in the original model formula.

coef.mvgam returns either summaries or full posterior estimates for GAM component coefficients

Author

Nicholas J Clark